requirements.txt
googlefinance==0.7
kafka-python==2.0.1
schedule==0.6.0
